+++ R/utils.R
@@ -1,6 +1,9 @@
 #' @importFrom magrittr %>%
 magrittr::`%>%`
 
+# Suppress R CMD check note
+#' @importFrom lifecycle deprecate_soft
+
 # Check that data is a mousetrap object
 is_mousetrap_data <- function(data){
 
@@ -15,8 +18,34 @@ extract_data <- function(data, use) {
   if (is_mousetrap_data(data)){
 
     extracted <- data[[use]]
+    
     if(is.null(extracted)){
-      stop("No data called '",use,"' found.")
+      
+      # Cover special cases due to the replacement of 
+      # mt_spatialize with mt_length_normalize
+      if (use == "ln_trajectories" & "sp_trajectories" %in% names(data)){
+        warning(
+          "Instead of ln_trajectories only sp_trajectories were found. ",
+          "These will be used. ",
+          "Please note that the mt_spatialize function has been deprecated ",
+          "and that the mt_length_normalize function should be used instead."
+          )
+        return(data[["sp_trajectories"]])
+        
+      } else if (use == "sp_trajectories" & "ln_trajectories" %in% names(data)){
+        warning(
+          "Instead of sp_trajectories only ln_trajectories were found. ",
+          "These will be used. ",
+          "Please note that since the mt_spatialize function has been replaced ",
+          "with the mt_length_normalize function, the resulting trajectories ",
+          "are now called ln_trajectories instead of sp_trajectories."
+        )
+        return(data[["ln_trajectories"]])
+        
+      } else {
+        stop("No data called '",use,"' found.")
+      }
+      
     }
     return(extracted)
 
@@ -31,6 +60,34 @@ extract_data <- function(data, use) {
   }
 
 }
+
+
+# Function for extracting dimensions from trajectory array
+# that also handles edge case of trajectory array with single trajectory
+extract_dimensions <- function(trajectories, dimensions){
+  
+  trajectories_dim <- dim(trajectories)
+  
+  # Standard case (no modification required)
+  if(trajectories_dim[1] != 1){
+    return(trajectories[,,dimensions])
+    
+  # Edge case of single trajectory
+  } else{
+    
+    # Use drop argument to prevent dropping of dimensions
+    trajectories <- trajectories[,,dimensions, drop = FALSE]
+    
+    # In case only a single dimension is extracted,
+    # drop third dimension from trajectory array
+    if(length(dimensions) == 1){
+      dim(trajectories) <- trajectories_dim[1:2]
+    }
+    
+    return(trajectories)
+  }
+}
+
 
 # Function to create results
 create_results <- function(data, results, use, save_as, ids=NULL, overwrite=TRUE) {